Add Password Recovery
Replace the 404.html page with a general error.html page that shows the
error code and its message.

Add a password recovery form to request a password reset.
The server requests an email address and generates a temporary password
reset token that lasts 5 minutes, using the itsdangerous'
URLSafeTimedSerializer.
Currently, the password reset URL is only available on the server
console.

Add a password reset function that receives a token generated by the
recover function.
With the GET method, the function checks if the specified token is
valid, in which case, the system renders a password reset page.
Otherwise, it returns a 401 HTTP Error.
With the POST method, the function gets the new passwords and checks if
they are valid, then updates the user's data on the database.